requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.26
matplotlib>=3.8
plotly>=5.20
requests>=2.31
python-dotenv>=1.0
rapidfuzz>=3.6
pytest>=8.0
ruff>=0.6
pillow>=10.0
